Improved efficiency and accuracy of the pathfinding algorithms.

+++ src/utils/pathfinding/DriveTask.java
@@ -2,19 +2,19 @@ package utils.pathfinding;
 
 public class DriveTask extends Task {
 	
-	private int distance;
+	private double distance;
 	private int speed;
 	
-	public DriveTask (int distance, int speed) {
+	public DriveTask (double distance, int speed) {
 		this.distance = distance;
 		this.speed = speed;
 	}
 	
-	public int getDistance() {
+	public double getDistance() {
 		return distance;
 	}
 	
-	public int getSpeed() {
+	public double getSpeed() {
 		return speed;
 	}
